refactored following buttons

+++ app/helpers/user_helper.rb
@@ -1,21 +1,20 @@
 module UserHelper
   def fullname(user)
-      user.first_name.capitalize + " " + user.last_name.capitalize
+    user.first_name.capitalize + " " + user.last_name.capitalize
   end
 
   def follow_button(user)
-    button_to 'Follow', follow_path(user)
+    unless is_me?(user)
+      if current_user.follow_user?(user)
+        button_to 'Unfollow', follow_path(user), method: :delete
+      else
+        button_to 'Follow', follow_path(user)
+      end
     end
-
-  def unfollow_button(user)
-    button_to 'Unfollow', follow_path(user), method: :delete
-  end
-
-  def follow_user?(user)
-    current_user.followed_users.include?(user)
   end
 
   def is_me?(user)
     current_user == user
   end
+
 end
